Move mechanics_ode and land model class to separate files in order to make mechanics_model a bit less complex
