+++ src/simcardems/mechanics_model.py
@@ -3,10 +3,11 @@ from enum import Enum
 
 import dolfin
 import pulse
-import ufl
 from mpi4py import MPI
 
 from . import utils
+from .newton_solver import MechanicsNewtonSolver_ODE
+
 
 logger = utils.getLogger(__name__)
 
@@ -14,236 +15,6 @@ logger = utils.getLogger(__name__)
 class BoundaryConditions(str, Enum):
     dirichlet = "dirichlet"
     rigid = "rigid"
-
-
-class Scheme(str, Enum):
-    fd = "fd"
-    bd = "bd"
-    analytic = "analytic"
-
-
-def _Zeta(Zeta_prev, A, c, dLambda, dt, scheme: Scheme):
-
-    if scheme == Scheme.analytic:
-        return Zeta_prev * dolfin.exp(-c * dt) + (A * dLambda / c * dt) * (
-            1 - dolfin.exp(-c * dt)
-        )
-
-    elif scheme == Scheme.bd:
-        return Zeta_prev + A * dLambda / (1 + c * dt)
-    else:
-        return Zeta_prev * (1 - c * dt) + A * dLambda
-
-
-class LandModel(pulse.ActiveModel):
-    def __init__(
-        self,
-        XS,
-        XW,
-        parameters,
-        mesh,
-        Zetas=None,
-        Zetaw=None,
-        lmbda=None,
-        f0=None,
-        s0=None,
-        n0=None,
-        eta=0,
-        scheme: Scheme = Scheme.analytic,
-        **kwargs,
-    ):
-        super().__init__(f0=f0, s0=s0, n0=n0)
-        self._eta = eta
-        self.function_space = pulse.QuadratureSpace(mesh, degree=3, dim=1)
-
-        self.XS = XS
-        self.XW = XW
-        self._parameters = parameters
-        self._t = 0.0
-        self._t_prev = 0.0
-        self._scheme = scheme
-
-        self._dLambda = dolfin.Function(self.function_space)
-        self.lmbda_prev = dolfin.Function(self.function_space)
-        if lmbda is not None:
-            self.lmbda_prev = lmbda
-        self.lmbda = dolfin.Function(self.function_space)
-
-        self._Zetas = dolfin.Function(self.function_space)
-        self.Zetas_prev = dolfin.Function(self.function_space)
-        if Zetas is not None:
-            self.Zetas_prev.assign(Zetas)
-
-        self._Zetaw = dolfin.Function(self.function_space)
-        self.Zetaw_prev = dolfin.Function(self.function_space)
-        if Zetaw is not None:
-            self.Zetaw_prev.assign(Zetaw)
-
-        self.V_cg1 = dolfin.FunctionSpace(mesh, "CG", 1)
-        self.Ta_current = dolfin.Function(self.function_space, name="Ta")
-        self.Ta_current_cg1 = dolfin.Function(self.V_cg1, name="Ta")
-
-    @property
-    def dLambda(self):
-        self._dLambda.vector()[:] = self.lmbda.vector() - self.lmbda_prev.vector()
-        return self._dLambda
-
-    @property
-    def Aw(self):
-        Tot_A = self._parameters["Tot_A"]
-        rs = self._parameters["rs"]
-        rw = self._parameters["rw"]
-        scale_popu_rw = self._parameters["scale_popu_rw"]
-        scale_popu_rs = self._parameters["scale_popu_rs"]
-        return (
-            Tot_A
-            * rs
-            * scale_popu_rs
-            / (rs * scale_popu_rs + rw * scale_popu_rw * (1 - (rs * scale_popu_rs)))
-        )
-
-    @property
-    def As(self):
-        return self.Aw
-
-    @property
-    def cw(self):
-        phi = self._parameters["phi"]
-        kuw = self._parameters["kuw"]
-        rw = self._parameters["rw"]
-
-        scale_popu_kuw = self._parameters["scale_popu_kuw"]
-        scale_popu_rw = self._parameters["scale_popu_rw"]
-        return (
-            kuw
-            * scale_popu_kuw
-            * phi
-            * (1 - (rw * scale_popu_rw))
-            / (rw * scale_popu_rw)
-        )
-
-    @property
-    def cs(self):
-        phi = self._parameters["phi"]
-        kws = self._parameters["kws"]
-        rs = self._parameters["rs"]
-        rw = self._parameters["rw"]
-        scale_popu_kws = self._parameters["scale_popu_kws"]
-        scale_popu_rw = self._parameters["scale_popu_rw"]
-        scale_popu_rs = self._parameters["scale_popu_rs"]
-        return (
-            kws
-            * scale_popu_kws
-            * phi
-            * rw
-            * scale_popu_rw
-            * (1 - (rs * scale_popu_rs))
-            / (rs * scale_popu_rs)
-        )
-
-    def update_Zetas(self):
-        self._Zetas.vector()[:] = _Zeta(
-            self.Zetas_prev.vector(),
-            self.As,
-            self.cs,
-            self.dLambda.vector(),
-            self.dt,
-            self._scheme,
-        )
-
-    @property
-    def Zetas(self):
-        return self._Zetas
-
-    def update_Zetaw(self):
-        self._Zetaw.vector()[:] = _Zeta(
-            self.Zetaw_prev.vector(),
-            self.Aw,
-            self.cw,
-            self.dLambda.vector(),
-            self.dt,
-            self._scheme,
-        )
-
-    @property
-    def Zetaw(self):
-        return self._Zetaw
-
-    @property
-    def dt(self) -> float:
-        from .setup_models import TimeStepper
-
-        return TimeStepper.ns2ms(self.t - self._t_prev)
-
-    @property
-    def t(self) -> float:
-        return self._t
-
-    def start_time(self, t):
-        self._t_prev = t
-        self._t = t
-
-    def update_time(self, t):
-        self._t_prev = self.t
-        self._t = t
-
-    def update_prev(self):
-        self.Zetas_prev.vector()[:] = self.Zetas.vector()
-        self.Zetaw_prev.vector()[:] = self.Zetaw.vector()
-        self.lmbda_prev.vector()[:] = self.lmbda.vector()
-        self.Ta_current.assign(
-            dolfin.project(
-                self.Ta,
-                self.function_space,
-                form_compiler_parameters={"representation": "quadrature"},
-            ),
-        )
-        utils.local_project(self.Ta_current, self.V_cg1, self.Ta_current_cg1)
-
-    @property
-    def Ta(self):
-        Tref = self._parameters["Tref"]
-        rs = self._parameters["rs"]
-        scale_popu_Tref = self._parameters["scale_popu_Tref"]
-        scale_popu_rs = self._parameters["scale_popu_rs"]
-        Beta0 = self._parameters["Beta0"]
-
-        _min = ufl.min_value
-        _max = ufl.max_value
-        if isinstance(self.lmbda, (int, float)):
-            _min = min
-            _max = max
-        lmbda = _min(1.2, self.lmbda)
-        h_lambda_prima = 1 + Beta0 * (lmbda + _min(lmbda, 0.87) - 1.87)
-        h_lambda = _max(0, h_lambda_prima)
-
-        return (
-            h_lambda
-            * (Tref * scale_popu_Tref / (rs * scale_popu_rs))
-            * (self.XS * (self.Zetas + 1) + self.XW * self.Zetaw)
-        )
-
-    def Wactive(self, F, **kwargs):
-        """Active stress energy"""
-
-        C = F.T * F
-        f = F * self.f0
-        self.lmbda.assign(
-            dolfin.project(
-                dolfin.sqrt(f**2),
-                self.function_space,
-                form_compiler_parameters={"representation": "quadrature"},
-            ),
-        )
-        self.update_Zetas()
-        self.update_Zetaw()
-
-        return pulse.material.active_model.Wactive_transversally(
-            Ta=self.Ta,
-            C=C,
-            f0=self.f0,
-            eta=self.eta,
-        )
 
 
 class ContinuationBasedMechanicsProblem(pulse.MechanicsProblem):
@@ -357,128 +128,6 @@ class MechanicsProblem(ContinuationBasedMechanicsProblem):
 
     def update_lmbda_prev(self):
         self.lmbda_prev.vector()[:] = self.lmbda.vector()
-
-
-class MechanicsNewtonSolver_ODE(dolfin.NewtonSolver):
-    def __init__(
-        self,
-        problem: pulse.NonlinearProblem,
-        state: dolfin.Function,
-        update_cb,
-        parameters=None,
-    ):
-        dolfin.PETScOptions.clear()
-
-        self._problem = problem
-        self._state = state
-        self._update_cb = update_cb
-        self._parameters = parameters
-
-        # Initializing Newton solver (parent class)
-        self.petsc_solver = dolfin.PETScKrylovSolver()
-        super().__init__(
-            self._state.function_space().mesh().mpi_comm(),
-            self.petsc_solver,
-            dolfin.PETScFactory.instance(),
-        )
-
-        # Setting default parameters
-        params = MechanicsNewtonSolver_ODE.default_solver_parameters()
-        for k, v in params.items():
-            if self.parameters.has_parameter(k):
-                self.parameters[k] = v
-            if self.parameters.has_parameter_set(k):
-                for subk, subv in params[k].items():
-                    self.parameters[k][subk] = subv
-        petsc = params.pop("petsc")
-        for k, v in petsc.items():
-            if v is not None:
-                dolfin.PETScOptions.set(k, v)
-        self.newton_verbose = params.pop("newton_verbose", False)
-        self.ksp_verbose = params.pop("ksp_verbose", False)
-        if self.newton_verbose:
-            dolfin.set_log_level(dolfin.LogLevel.INFO)
-            self.parameters["report"] = True
-        if self.ksp_verbose:
-            self.parameters["lu_solver"]["report"] = True
-            self.parameters["lu_solver"]["verbose"] = True
-            self.parameters["krylov_solver"]["monitor_convergence"] = True
-            dolfin.PETScOptions.set("ksp_monitor_true_residual")
-        self.linear_solver().set_from_options()
-
-    @staticmethod
-    def default_solver_parameters():
-        return {
-            "petsc": {
-                # "ksp_type": "preonly",
-                "ksp_type": "gmres",
-                "pc_type": "lu",
-                "pc_factor_mat_solver_type": "mumps",
-                "mat_mumps_icntl_33": 0,
-            },
-            "newton_verbose": False,
-            "ksp_verbose": False,
-            # "linear_solver": "mumps",
-            "linear_solver": "gmres",
-            "error_on_nonconvergence": True,
-            "relative_tolerance": 1e-5,
-            "absolute_tolerance": 1e-5,
-            "maximum_iterations": 20,
-            "report": False,
-            # },
-            "krylov_solver": {
-                "absolute_tolerance": 1e-10,
-                "relative_tolerance": 1e-10,
-                "maximum_iterations": 1000,
-                "monitor_convergence": False,
-            },
-            "lu_solver": {"report": False, "symmetric": False, "verbose": False},
-        }
-
-    def converged(self, r, p, i):
-        self._converged_called = True
-
-        if dolfin.MPI.rank(dolfin.MPI.comm_world) == 0:
-            # Print all residuals
-            residual = r.norm("l2")
-            with open("residual.txt", "a") as rfile:
-                rfile.write(str(residual) + "\t")
-
-        return super().converged(r, p, i)
-
-    def solver_setup(self, A, J, p, i):
-        self._solver_setup_called = True
-        super().solver_setup(A, J, p, i)
-
-    def update_solution(self, x, dx, rp, p, i):
-        self._update_solution_called = True
-
-        # Update x from the dx obtained from linear solver (Newton iteration) :
-        # x = -rp*dx (rp : relax param)
-        super().update_solution(x, dx, rp, p, i)
-
-        # Updating form of MechanicsProblem (from current lmbda, zetas, zetaw, ...)
-        self._state.vector().set_local(x)
-        # self._mech_problem._init_forms()
-        # Recompute Zetas, Zetaw, Ta, lmbda
-        # self._mech_problem.material.active.update_prev()
-        self._update_cb()
-        # Re-init this solver with the new problem (note : done in _init_forms)
-        # self.__init__(self._mech_problem)
-
-    def solve(self):
-        self._solve_called = True
-        ret = super().solve(self._problem, self._state.vector())
-        self._state.vector().apply("insert")
-        return ret
-
-    # DEBUGGING
-    # This is just to check if we are using the overloaded functions
-    def check_overloads_called(self):
-        assert getattr(self, "_converged_called", False)
-        assert getattr(self, "_solver_setup_called", False)
-        assert getattr(self, "_update_solution_called", False)
-        assert getattr(self, "_solve_called", False)
 
 
 class RigidMotionProblem(MechanicsProblem):
